+++ packages/cuaktask/src/task/modules/DependentTaskBuildOperation.spec.ts
@@ -40,6 +40,11 @@ describe(DependentTaskBuilder.name, () => {
           dependencies: [],
           kind: taskKindFixture,
           perform: jest.fn(),
+          result: {
+            get: () => {
+              throw new Error();
+            },
+          },
           status: TaskStatus.NotStarted,
         };
 
@@ -100,6 +105,11 @@ describe(DependentTaskBuilder.name, () => {
         dependencies: [],
         kind: dependentTaskKindFixture,
         perform: jest.fn(),
+        result: {
+          get: () => {
+            throw new Error();
+          },
+        },
         status: TaskStatus.NotStarted,
       };
       taskKindFixture = 'some-kind';
@@ -107,6 +117,11 @@ describe(DependentTaskBuilder.name, () => {
         dependencies: [],
         kind: taskKindFixture,
         perform: jest.fn(),
+        result: {
+          get: () => {
+            throw new Error();
+          },
+        },
         status: TaskStatus.NotStarted,
       };