+++ packages/cuaktask/src/task/modules/DependentTaskBuilder.spec.ts
@@ -78,6 +78,11 @@ describe(DependentTaskBuilder.name, () => {
           dependencies: [],
           kind: taskKindFixture,
           perform: jest.fn(),
+          result: {
+            get: () => {
+              throw new Error();
+            },
+          },
           status: TaskStatus.NotStarted,
         };