+++ packages/iocuak/src/container/services/cuaktask/ContainerInstanceServiceImplementation.spec.ts
@@ -64,6 +64,11 @@ describe(ContainerInstanceServiceImplementation.name, () => {
         dependencies: [],
         kind: taskKindFixture,
         perform: jest.fn(),
+        result: {
+          get: () => {
+            throw new Error();
+          },
+        },
         status: TaskStatus.NotStarted,
       };